mobile/android/app/src/main/python/requirements.txt
pillow==11.2.1
tensorflow==2.10.0
rembg==2.0.61
tqdm
